remove redundant casts

+++ src/main/java/com/darkona/adventurebackpack/entity/EntityInflatableBoat.java
@@ -123,7 +123,7 @@ public class EntityInflatableBoat extends EntityBoat implements /* IInventoryTan
             d4 = Math.sin((double) this.rotationYaw * Math.PI / 180.0D);
 
             for (j = 0; (double) j < 1.0D + horizontalMotion * 60.0D; ++j) {
-                double d5 = (double) (this.rand.nextFloat() * 2.0F - 1.0F);
+                double d5 = this.rand.nextFloat() * 2.0F - 1.0F;
                 double d6 = (double) (this.rand.nextInt(2) * 2 - 1) * 0.7D;
                 double d8;
                 double d9;
@@ -200,10 +200,10 @@ public class EntityInflatableBoat extends EntityBoat implements /* IInventoryTan
             if (this.riddenByEntity != null && this.riddenByEntity instanceof EntityLivingBase) {
                 EntityLivingBase entitylivingbase = (EntityLivingBase) this.riddenByEntity;
                 float f = this.riddenByEntity.rotationYaw + -entitylivingbase.moveStrafing * 90.0F;
-                this.motionX += -Math.sin((double) (f * (float) Math.PI / 180.0F)) * this.speedMultiplier
+                this.motionX += -Math.sin(f * (float) Math.PI / 180.0F) * this.speedMultiplier
                         * (double) entitylivingbase.moveForward
                         * 0.05000000074505806D;
-                this.motionZ += Math.cos((double) (f * (float) Math.PI / 180.0F)) * this.speedMultiplier
+                this.motionZ += Math.cos(f * (float) Math.PI / 180.0F) * this.speedMultiplier
                         * (double) entitylivingbase.moveForward
                         * 0.05000000074505806D;
             }
@@ -267,12 +267,12 @@ public class EntityInflatableBoat extends EntityBoat implements /* IInventoryTan
              */
 
             this.rotationPitch = 0.0F;
-            d4 = (double) this.rotationYaw;
+            d4 = this.rotationYaw;
             d11 = this.prevPosX - this.posX;
             d12 = this.prevPosZ - this.posZ;
 
             if (d11 * d11 + d12 * d12 > 0.001D) {
-                d4 = (double) ((float) (Math.atan2(d12, d11) * 180.0D / Math.PI));
+                d4 = (float) (Math.atan2(d12, d11) * 180.0D / Math.PI);
             }
 
             double d7 = MathHelper.wrapAngleTo180_double(d4 - (double) this.rotationYaw);
